refactor(ai-usage): split 1200-line page into focused modules
No behavior change. Extract the single-file AI Usage & Budget page
into components/ai-usage/ to keep cohesive chunks loadable on their
own instead of forcing the full mental context on any change.

- metrics.ts — pure types/logic: BudgetMetrics, computeMetrics,
  statusStyles, MS_PER_DAY, formatCurrency, formatShortDate
- primitives.tsx — shared UI: SectionLabel, Stat (info tooltip)
- trajectory-chart.tsx — TrajectoryChart + local ChartLegendItem
- budget-health-hero.tsx — BudgetHealthHero (headline, 4 stats, chart)
- usage-breakdown.tsx — UsageSplitCard, TopSpendersCard
- budget-controls-panel.tsx — BudgetControlsPanel (inputs, alerts, guardrails)

The page file is now 161 lines of orchestration (was 1149) — data
fetching, optimistic metrics computation, and layout. Follows the
same flat-folder pattern used by components/workspace-settings/.
